clean up duplicate and unused code
