+++ services/QuillLMS/client/app/bundles/Teacher/components/progress_reports/diagnostic_reports/diagnostics/recommendationsTable.tsx
@@ -5,7 +5,6 @@ import {
   noDataYet,
   recommendedGlyph,
   correctImage,
-  releaseMethodToDisplayName,
   baseDiagnosticImageSrc,
   DEFAULT_LEFT_PADDING,
   MOBILE_WIDTH,
@@ -16,19 +15,16 @@ import {
   Student,
 } from './interfaces'
 import StudentNameOrTooltip from './studentNameOrTooltip'
+import ActivityPackHeader from './activityPackHeader'
 
 import {
-  helpIcon,
-  Tooltip,
   smallWhiteCheckIcon,
 } from '../../../../../Shared/index'
 import useWindowSize from '../../../../../Shared/hooks/useWindowSize'
 
 const openInNewTabIcon = <img alt="Open in new tab icon" src={`${baseDiagnosticImageSrc}/icons-open-in-new.svg`} />
-const ellipsesIcon = <img alt="Open menu icon" src={`${baseDiagnosticImageSrc}/ellipses_icon.svg`} />
 
 const TABLE_LEFT_PADDING = 0
-const TABLE_RIGHT_PADDING = 34
 
 interface RecommendationsTableProps {
   recommendations: Recommendation[];
@@ -55,35 +51,6 @@ interface RecommendationCellProps {
   selectionIndex: number;
   completedCount?: number;
   activityCount?: number;
-}
-
-interface StickyTableStyle {
-  position: string,
-  top: number,
-  left: number,
-  right: number,
-  zIndex: number,
-  minWidth: string,
-  width: string,
-  overflowX: string
-}
-
-const ActivityPackHeader = ({ name, activityPackId, activityCount, handleSelectAllClick, style, }) => {
-  return (
-    <th className="recommendation-header" key={name} style={style || {}}>
-      <div className="name-and-tooltip">
-        <span>{name}</span>
-        <a aria-label="Preview the activity pack" href={`/activities/packs/${activityPackId}`} rel="noopener noreferrer" target="_blank"><img alt="" src={helpIcon.src} /></a>
-      </div>
-      <div className="activity-count-and-select-all">
-        <span className="activity-count">{activityCount} {activityCount === 1 ? 'activity' : 'activities'}</span>
-        <Tooltip
-          tooltipText="Select all column"
-          tooltipTriggerText={<button aria-label="Select all column" className="interactive-wrapper" onClick={handleSelectAllClick} type="button">{ellipsesIcon}</button>}
-        />
-      </div>
-    </th>
-  )
 }
 
 const RecommendationCell = ({ student, isAssigned, isRecommended, isSelected, setSelections, selections, selectionIndex, activityCount, completedCount, }: RecommendationCellProps) => {
@@ -175,54 +142,13 @@ const StudentRow = ({ student, selections, recommendations, previouslyAssignedRe
 
 const RecommendationsTable = ({ recommendations, responsesLink, students, selections, previouslyAssignedRecommendations, setSelections, studentsWhoCompletedDiagnostic, studentsWhoCompletedAssignedRecommendations, postDiagnosticUnitTemplateId, setPostTestSelections, postTestSelections, previouslyAssignedPostTestStudentIds, }: RecommendationsTableProps) => {
   const size = useWindowSize();
-
-  const [isSticky, setIsSticky] = React.useState(false);
   const recommendationsTableRef = React.useRef(null);
-  const [stickyTableStyle, setStickyTableStyle] = React.useState<StickyTableStyle>({
-    position: "fixed",
-    top: 0,
-    left: 0,
-    right: 0,
-    zIndex: 3,
-    width: "63%",
-    minWidth: "63%",
-    overflowX: "scroll",
-  })
-  const [stickyTableRightOffset, setStickyTableRightOffset] = React.useState(0)
 
   function paddingLeft() {
     if (MOBILE_WIDTH >= window.innerWidth) { return DEFAULT_LEFT_PADDING_FOR_MOBILE }
     const explanation = document.getElementsByClassName('explanation')[0]
     return explanation && DEFAULT_LEFT_PADDING
   }
-
-  const handleScroll = React.useCallback(({ top, bottom, left, right, }) => {
-    if (top <= 0 && bottom > 92) {
-      setStickyTableStyle(oldStickyTableStyle => ({ ...oldStickyTableStyle, left: left }))
-      setStickyTableRightOffset(right)
-      !isSticky && setIsSticky(true);
-    } else {
-      isSticky && setIsSticky(false);
-    }
-  }, [isSticky]);
-
-  const onScroll = () => {
-    if (recommendationsTableRef && recommendationsTableRef.current) {
-      handleScroll(recommendationsTableRef.current.getBoundingClientRect());
-    }
-  }
-
-  React.useEffect(() => {
-    onScroll()
-  }, [size])
-
-  React.useEffect(() => {
-    window.addEventListener("scroll", onScroll);
-
-    return () => {
-      window.removeEventListener("scroll", onScroll);
-    };
-  }, [handleScroll]);
 
   function onSelectAllColumnClick(activityPackId) {
     const newSelections = [...selections];
@@ -232,11 +158,6 @@ const RecommendationsTable = ({ recommendations, responsesLink, students, select
     newSelections[selectionIndex].students = studentIdsWhoHaveCompletedDiagnostic
 
     setSelections(newSelections)
-  }
-
-  function onSelectAllForPostDiagnosticClick() {
-    const studentIdsWhoHaveCompletedDiagnostic = students.filter(s => s.completed).map(s => s.id)
-    setPostTestSelections(studentIdsWhoHaveCompletedDiagnostic)
   }
 
   const tableHeaders = recommendations && recommendations.map(recommendation => {
@@ -283,62 +204,6 @@ const RecommendationsTable = ({ recommendations, responsesLink, students, select
           {tableHeaders}
         </tr>
       </thead>
-    )
-  }
-
-  const renderPostTestTableHeader = (sticky) => {
-    const style = sticky ? { position: 'inherit' } : {}
-
-    const anchorElement = document.getElementsByClassName('corner-header')[0]
-
-    if (anchorElement) {
-      style.height = anchorElement.getBoundingClientRect().height
-    }
-
-    return (
-      <ActivityPackHeader
-        activityCount={1}
-        activityPackId={postDiagnosticUnitTemplateId}
-        handleSelectAllClick={onSelectAllForPostDiagnosticClick}
-        name="Diagnostic (Post) Test"
-        style={style}
-      />
-    )
-  }
-
-  const renderStickyRecommendationsTable = () => {
-    // an arbitrary, non-resizing element that is the same width that we need this table to be
-    const anchorElement = document.getElementsByClassName('independent-practice-recommendations-buttons')[0]
-
-    if (!(isSticky && tableHasContent && anchorElement)) { return }
-
-    const width = anchorElement.getBoundingClientRect().width + TABLE_RIGHT_PADDING
-
-    return (
-      <table
-        className={`${recommendationsTableClassName} sticky`}
-        style={{ ...stickyTableStyle }}
-      >
-        {renderRecommendationsTableHeader(true)}
-      </table>
-    )
-  }
-
-  const renderStickyPostTestTable = () => {
-    // an arbitrary, non-resizing element that is the same width that we need this table to be
-    const widthAnchorElement = document.getElementsByClassName('post-test-assignment-button')[0]
-
-    if (!(isSticky && tableHasContent && widthAnchorElement)) { return }
-
-    const width = widthAnchorElement.getBoundingClientRect().width
-
-    return (
-      <table
-        className={`${recommendationsTableClassName} sticky`}
-        style={{ ...stickyTableStyle, left: stickyTableRightOffset, minWidth: width, width }}
-      >
-        {renderPostTestTableHeader(true)}
-      </table>
     )
   }
